requirements.txt
requests==2.32.5
python-dotenv==1.2.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.4.0
